examples/enterprise-security/testing-app/requirements.txt
flask==3.0.0
redis==5.0.1
gunicorn==21.2.0
python-dotenv==1.0.0
opencensus-ext-azure==1.1.13
opencensus-ext-flask==0.8.1
opencensus-ext-logging==0.1.1
Werkzeug==3.0.1
